lundi 7 juin 42

+++ ft_atoi.c
@@ -3,26 +3,26 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kyoko <kyoko@student.42.fr>                +#+  +:+       +#+        */
+/*   By: mchampag <mchampag@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/12 11:25:00 by mchampag          #+#    #+#             */
-/*   Updated: 2021/06/06 00:20:55 by kyoko            ###   ########.fr       */
+/*   Updated: 2021/06/07 10:47:24 by mchampag         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 /*
 ** atoi() ASCII to integer
-** 	Convertir une chaîne de caractères (48 à 57) en entier, prend en compte 
+** 	Convertir une chaîne de caractères (48 à 57) en entier, prend en compte
 **	le signe (positif '+' et négatif '-').
 ** PARAMÈTRE
-**	str: contient la chaîne de caractères à convertir en une donnée de type 
+**	str: contient la chaîne de caractères à convertir en une donnée de type
 **			int.
 ** VALEUR RENVOYÉE
-** 	Le résultat de la conversion ou 0 si la chaîne ne contient pas une donnée 
+** 	Le résultat de la conversion ou 0 si la chaîne ne contient pas une donnée
 **	numérique entière.
 ** ATTENTION
-** 	Il n'est pas possible de distinguer la chaîne "0" d'une chaîne 
-** 	ne contenant pas un nombre entier. 
+** 	Il n'est pas possible de distinguer la chaîne "0" d'une chaîne
+** 	ne contenant pas un nombre entier.
 ** CARACTÈRES À NE PAS TENIR EN COMPTE
 ** \a : appel (cloche) '7'
 ** \b : retour arrière '8'
@@ -32,15 +32,15 @@
 ** \f : changement de page '12'
 ** \r : retour chariot '13'
 ** \e ou ' ' : expace '27'
-** 
+**
 */
 
 #include "libft.h"
-   
+
 int		ft_atoi(const char *str)
 {
-	int	sign;
-	int	converted;
+	long	sign;
+	long	converted;
 
 	sign = 1;
 	converted = 0;
